Replace CasesTable interface with full class
