+++ jgiven-core/src/main/java/com/tngtech/jgiven/report/CasesTable.java
@@ -1,47 +1,92 @@
 package com.tngtech.jgiven.report;
 
 import com.tngtech.jgiven.report.model.ExecutionStatus;
+import com.tngtech.jgiven.report.model.ScenarioCaseModel;
+import com.tngtech.jgiven.report.model.ScenarioModel;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
  * Represents multiple scenario cases as a table structure for simpler reporting.
  */
-public interface CasesTable {
+public class CasesTable {
+    private final ScenarioModel scenarioModel;
+
+    public CasesTable(ScenarioModel scenarioModel) {
+        this.scenarioModel = scenarioModel;
+    }
+
     /**
      * The placeholders of the data table.
      */
-    List<String> placeholders();
+    public List<String> placeholders() {
+        List<String> placeHoldersList = new ArrayList<>(scenarioModel.getDerivedParameters());
+        List<ScenarioCaseModel> scenarioCases = scenarioModel.getScenarioCases();
+        if (!scenarioCases.isEmpty() && scenarioCases.get(0).hasDescription()) {
+            placeHoldersList.add(0, "Description");
+        }
+        return placeHoldersList;
+    }
 
     /**
      * The rows of the table, not including the header.
      */
-    List<Row> rows();
+    public List<Row> rows() {
+        List<Row> rows = new ArrayList<>();
+        for (ScenarioCaseModel caseModel : scenarioModel.getScenarioCases()) {
+            rows.add(new Row(caseModel));
+        }
+        return rows;
+    }
 
     /**
      * Represents one case of a scenario.
      */
-    interface Row {
+    public static class Row {
+
+        private final ScenarioCaseModel caseModel;
+
+        public Row(ScenarioCaseModel caseModel) {
+            this.caseModel = caseModel;
+        }
+
         /**
          * The row number starting from 1.
          */
-        int nr();
+        public int nr() {
+            return caseModel.getCaseNr();
+        }
 
         /**
          * The execution status of the case.
          */
-        ExecutionStatus status();
+        public ExecutionStatus status() {
+            return caseModel.getExecutionStatus();
+        }
 
         /**
          * The argument values of the case.
          */
-        List<String> arguments();
+        public List<String> arguments() {
+            List<String> arguments = new ArrayList<>(caseModel.getDerivedArguments());
+            if (caseModel.hasDescription()) {
+                arguments.add(0, caseModel.getDescription());
+            }
+            return arguments;
+        }
 
-        /** The error message of the case if any.
+        /**
+         *  The error message of the case if any.
          */
-        String errorMessage();
+        public String errorMessage() {
+            return caseModel.getErrorMessage();
+        }
 
-        /** The stack trace of the case if any.
+        /**
+         *  The stack trace of the case if any.
          */
-        List<String> stackTrace();
+        public List<String> stackTrace() {
+            return caseModel.getStackTrace();
+        }
     }
 }
